+++ src/main/kotlin/io/snyk/plugin/ui/jcef/Utils.kt
@@ -6,10 +6,12 @@ import com.intellij.ui.jcef.JBCefBrowserBuilder
 import org.cef.handler.CefLoadHandlerAdapter
 import java.awt.Component
 
+typealias LoadHandlerGenerator = (jbCefBrowser: JBCefBrowser) -> CefLoadHandlerAdapter
+
 object JCEFUtils {
     fun getJBCefBrowserComponentIfSupported(
         html: String,
-        loadHandlerGenerator: (jbCefBrowser: JBCefBrowser) -> CefLoadHandlerAdapter,
+        loadHandlerGenerators: List<LoadHandlerGenerator>,
     ): Component? {
         if (!JBCefApp.isSupported()) {
             return null
@@ -21,9 +23,10 @@ object JCEFUtils {
                 .setMouseWheelEventEnable(true).build()
         jbCefBrowser.setOpenLinksInExternalBrowser(true)
 
-        val loadHandler = loadHandlerGenerator(jbCefBrowser)
-        cefClient.addLoadHandler(loadHandler, jbCefBrowser.cefBrowser)
-
+        for (loadHandlerGenerator in loadHandlerGenerators) {
+            val loadHandler = loadHandlerGenerator(jbCefBrowser)
+            cefClient.addLoadHandler(loadHandler, jbCefBrowser.cefBrowser)
+        }
         jbCefBrowser.loadHTML(html, jbCefBrowser.cefBrowser.url)
 
         return jbCefBrowser.component