+++ src/test/kotlin/io/snyk/plugin/ui/toolwindow/SuggestionDescriptionPanelFromLSCodeTest.kt
@@ -12,7 +12,6 @@ import io.mockk.mockkObject
 import io.mockk.unmockkAll
 import io.snyk.plugin.Severity
 import io.snyk.plugin.SnykFile
-import io.snyk.plugin.pluginSettings
 import io.snyk.plugin.resetSettings
 import io.snyk.plugin.ui.jcef.JCEFUtils
 import io.snyk.plugin.ui.toolwindow.panels.SuggestionDescriptionPanelFromLS
@@ -82,7 +81,9 @@ class SuggestionDescriptionPanelFromLSCodeTest : BasePlatformTestCase() {
     }
 
     @Test
-    fun `test createUI should build the right panels for Snyk Code`() {
+    fun `test createUI should build the right panels for Snyk Code if HTML is not allowed`() {
+        every { issue.canLoadSuggestionPanelFromHTML() } returns false
+
         cut = SuggestionDescriptionPanelFromLS(snykFile, issue)
 
         val issueNaming = getJLabelByText(cut, issue.issueNaming())
@@ -108,22 +109,9 @@ class SuggestionDescriptionPanelFromLSCodeTest : BasePlatformTestCase() {
     }
 
     @Test
-    fun `test createUI should build panel with issue message as overview label if the feature flag is not enabled`() {
-        every { issue.details() } returns ""
-        cut = SuggestionDescriptionPanelFromLS(snykFile, issue)
-
-        val actual = getJLabelByText(cut, "<html>Test message</html>")
-        assertNotNull(actual)
-
-        val actualBrowser = getJBCEFBrowser(cut)
-        assertNull(actualBrowser)
-    }
-
-    @Test
-    fun `test createUI should build panel with issue message as overview label if HTML is not allowed, even if the feature flag is enabled`() {
-        pluginSettings().isGlobalIgnoresFeatureEnabled = true
-
+    fun `test createUI should build panel with issue message as overview label if HTML is not allowed`() {
         every { issue.canLoadSuggestionPanelFromHTML() } returns false
+
         cut = SuggestionDescriptionPanelFromLS(snykFile, issue)
 
         val actual = getJLabelByText(cut, "<html>Test message</html>")
@@ -134,9 +122,7 @@ class SuggestionDescriptionPanelFromLSCodeTest : BasePlatformTestCase() {
     }
 
     @Test
-    fun `test createUI should show nothing if feature flag is enabled but JCEF is not`() {
-        pluginSettings().isGlobalIgnoresFeatureEnabled = true
-
+    fun `test createUI should show nothing if HTML is allowed but JCEF is not supported`() {
         mockkObject(JCEFUtils)
         every { JCEFUtils.getJBCefBrowserComponentIfSupported(eq("<html>HTML message</html>"), any()) } returns null
 
@@ -152,9 +138,7 @@ class SuggestionDescriptionPanelFromLSCodeTest : BasePlatformTestCase() {
     }
 
     @Test
-    fun `test createUI should build panel with HTML from details if feature flag is enabled`() {
-        pluginSettings().isGlobalIgnoresFeatureEnabled = true
-
+    fun `test createUI should build panel with HTML from details if allowed`() {
         val mockJBCefBrowserComponent = JLabel("<html>HTML message</html>")
         mockkObject(JCEFUtils)
         every {
@@ -173,9 +157,7 @@ class SuggestionDescriptionPanelFromLSCodeTest : BasePlatformTestCase() {
     }
 
     @Test
-    fun `test getStyledHTML should inject CSS into the HTML`() {
-        pluginSettings().isGlobalIgnoresFeatureEnabled = true
-
+    fun `test getStyledHTML should inject CSS into the HTML if allowed`() {
         every { issue.details() } returns "<html><head>\${ideStyle}</head>HTML message</html>"
         every { issue.canLoadSuggestionPanelFromHTML() } returns true
         cut = SuggestionDescriptionPanelFromLS(snykFile, issue)